+++ functions.php
@@ -29,6 +29,8 @@ require get_template_directory() . '/inc/theme-admin-settings.php';
 
 require get_template_directory() . '/inc/theme-template-tags.php';
 
+require get_template_directory() . '/inc/rest-job-filters.php';
+
 require get_template_directory() . '/inc/helpers.php';
 
 require get_template_directory() . '/inc/performance.php';